Create README.md
There has always been a major problem of communication between special people and normal people. Sign Language Converter is a program developed to detect the hand signs and then convert them into short English sentences as well as it can convert the text into sound which also helps the user to understand the text in a much better way. The main purpose or the background of this program is to help deaf and dumb people. It can help them to understand the text written on the screen just by making different hand gestures. By using different hand conditions a whole language for those people can be developed and serve humanity.
In this project, we will use an efficient and effective method for hand gesture recognition. The hand region is detected through the background subtraction method. Then, the palm and fingers are split so as to recognize the fingers. After the fingers are recognized, the hand gesture can be classified through a simple rule classifier.